src/main/java/com/example/mercadonabackend/api/auth/UserController.java
package com.example.mercadonabackend.api.auth;

import com.example.mercadonabackend.Service.UserService;
import com.example.mercadonabackend.dto.RegistrationDto;
import com.example.mercadonabackend.pojo.UserEntity;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Controller;
import org.springframework.ui.Model;
import org.springframework.validation.BindingResult;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.ModelAttribute;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.servlet.ModelAndView;

@Controller
@RequiredArgsConstructor
public class UserController {

    private final UserService userService;

    @GetMapping("/login")
    public String loginPage() {
        return "login";
    }


    @GetMapping("/logout")
    public String logout() {
        return "index";
    }


}
